+++ idea/testData/intentions/convertFunctionTypeParameterToReceiver/notOnFunctionTypeParameter.kt
@@ -1,0 +1,4 @@
+// IS_APPLICABLE: false
+fun foo(f: (Int, Boolean) -> <caret>String) {
+
+}